Method confirmGeneral does not return false on OkResult -> fixed
The implementation of confirmGeneral(data: MessageParameter) returns true in case that the result of the dialogRef is equal to the YesResult. But within the MessageParameter you can decide to work with the OK & Cancel-Buttons.
So I would expect that the OkResult is also a positive result of the dialogReg, and the return value should be true.